+++ src/main/java/com/ycloud/client/api/WhatsappTemplatesApi.java
@@ -798,7 +798,7 @@ public class WhatsappTemplatesApi {
 
         /**
          * Set filterWabaId
-         * @param filterWabaId WhatsApp Business Account ID. (optional)
+         * @param filterWabaId **Required**. WhatsApp Business Account ID. (optional)
          * @return ApiListRequest
          */
         public ApiListRequest filterWabaId(String filterWabaId) {