+++ src/main/java/com/ycloud/client/model/Contact.java
@@ -124,6 +124,13 @@ public class Contact {
   @SerializedName(SERIALIZED_NAME_CUSTOM_ATTRIBUTES)
   private List<ContactCustomAttribute> customAttributes = null;
 
+  public static final String SERIALIZED_NAME_OWNER_EMAIL = "ownerEmail";
+  /**
+   * The email address of the contact&#39;s owner.
+  **/
+  @SerializedName(SERIALIZED_NAME_OWNER_EMAIL)
+  private String ownerEmail;
+
   public Contact() { 
   }
 
@@ -199,7 +206,7 @@ public class Contact {
   /**
    * Two-letter country abbreviation. See [ISO 3166-1 alpha-2 country code](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2).
    **/
-  @ApiModelProperty(example = "GB", value = "Two-letter country abbreviation. See [ISO 3166-1 alpha-2 country code](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2).")
+  @ApiModelProperty(example = "US", value = "Two-letter country abbreviation. See [ISO 3166-1 alpha-2 country code](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2).")
   public void setCountryCode(String countryCode) {
     this.countryCode = countryCode;
   }
@@ -402,6 +409,32 @@ public class Contact {
     this.customAttributes = customAttributes;
   }
 
+
+  public Contact ownerEmail(String ownerEmail) {
+    
+    this.ownerEmail = ownerEmail;
+    return this;
+  }
+
+   /**
+   * The email address of the contact&#39;s owner.
+   * @return ownerEmail
+  **/
+  @javax.annotation.Nullable
+
+  public String getOwnerEmail() {
+    return ownerEmail;
+  }
+
+
+  /**
+   * The email address of the contact&#39;s owner.
+   **/
+  @ApiModelProperty(example = "support@example.com", value = "The email address of the contact's owner.")
+  public void setOwnerEmail(String ownerEmail) {
+    this.ownerEmail = ownerEmail;
+  }
+
   /**
    * A container for additional, undeclared properties.
    * This is a holder for any undeclared properties as specified with
@@ -457,13 +490,14 @@ public class Contact {
         Objects.equals(this.lastSeen, contact.lastSeen) &&
         Objects.equals(this.tags, contact.tags) &&
         Objects.equals(this.createTime, contact.createTime) &&
-        Objects.equals(this.customAttributes, contact.customAttributes)&&
+        Objects.equals(this.customAttributes, contact.customAttributes) &&
+        Objects.equals(this.ownerEmail, contact.ownerEmail)&&
         Objects.equals(this.additionalProperties, contact.additionalProperties);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(id, nickname, countryCode, countryName, phoneNumber, email, lastSeen, tags, createTime, customAttributes, additionalProperties);
+    return Objects.hash(id, nickname, countryCode, countryName, phoneNumber, email, lastSeen, tags, createTime, customAttributes, ownerEmail, additionalProperties);
   }
 
   @Override
@@ -480,6 +514,7 @@ public class Contact {
     sb.append("    tags: ").append(toIndentedString(tags)).append("\n");
     sb.append("    createTime: ").append(toIndentedString(createTime)).append("\n");
     sb.append("    customAttributes: ").append(toIndentedString(customAttributes)).append("\n");
+    sb.append("    ownerEmail: ").append(toIndentedString(ownerEmail)).append("\n");
     sb.append("    additionalProperties: ").append(toIndentedString(additionalProperties)).append("\n");
     sb.append("}");
     return sb.toString();
@@ -513,6 +548,7 @@ public class Contact {
     openapiFields.add("tags");
     openapiFields.add("createTime");
     openapiFields.add("customAttributes");
+    openapiFields.add("ownerEmail");
 
     // a set of required properties/fields (JSON key names)
     openapiRequiredFields = new HashSet<String>();
@@ -573,6 +609,9 @@ public class Contact {
         for (int i = 0; i < jsonArraycustomAttributes.size(); i++) {
           ContactCustomAttribute.validateJsonObject(jsonArraycustomAttributes.get(i).getAsJsonObject());
         };
+      }
+      if (jsonObj.get("ownerEmail") != null && !jsonObj.get("ownerEmail").isJsonPrimitive()) {
+        throw new IllegalArgumentException(String.format("Expected the field `ownerEmail` to be a primitive type in the JSON string but got `%s`", jsonObj.get("ownerEmail").toString()));
       }
   }